Undo fixes and avoid clearing selections on undo
* Avoid a compiler warning about comparing signed/unsigned int in "undomanager.cpp".
* Add an option to `mainWindow_c::ReplacePuzzle` used by undo/redo to keep previous selections in the UI.
* Fix a handfull of places we recorded undo states before actually performing changes to the puzzle.

+++ src/gui/undomanager.cpp
@@ -64,7 +64,7 @@ puzzle_c * UndoManager_c::undo(void) {
 }
 
 puzzle_c * UndoManager_c::redo(void) {
-  if (currentState >= states.size()-1) { return NULL; }
+  if ((unsigned int) currentState >= states.size()-1) { return NULL; }
   currentState++;
 
   std::istringstream stateString(states[currentState]);
